+++ src/Neurons/InputNeuron.java
@@ -4,19 +4,19 @@ package Neurons;
  * Created by qadirhaqq on 4/1/16.
  */
 public class InputNeuron implements Neuron{
-    private float value;
+    private double value;
     /**
      * The input neuron to compute
      * @param value
      */
-    public InputNeuron(float value){
+    public InputNeuron(double value){
         this.value = value;
     }
-    public void setValue(float value){
+    public void setValue(double value){
         this.value = value;
     }
     @Override
-    public float compute() {
+    public double compute() {
         return value;
     }
 }
