+++ src/mathematics/Function.java
@@ -6,5 +6,5 @@ import java.util.ArrayList;
  * Created by qadirhaqq on 4/1/16.
  */
 public interface Function {
-    public float calculate(Float in);
+    public double calculate(double in);
 }
